Hooked up webhooks API requests, Updated S4 dependency with forked version
S4 master branch has a change which fixes a compile error (preventing BuddyBuild from building!). I've forked off master and created a new release which should fix this issue.

Also in this branch I've added the code necessary for the webhook methods to become active and make the actual requests