Removed S4 Dependency, Replaced with Vapor
From the get go my goal with this project was to provide a small framework that could be used within my projects (using Vapor) to act as a middleman for the Monzo API. The S4 standards stood out as something that quite a few server-side Swift providers were using, but it appears as of Version 2 of Vapor this is not 100% the case and have had quite a few difficulties getting it to work nicely together.

With that in mind I've decided to replace the S4 dependency with Vapor directly, meaning I drop support for other providers - but it does open the door for additional functionality like better middleware/routing for authenticating.

Whilst this does build fine, I'm waiting to test it in a different project before updating all the logic to take better use of the Vapor capabilities