Add convenience extension to assert JSON values
Prefer using a throw over the many force unwraps! Every value type will need to have it's type pasted in, but not many different types are supported in the API and they don't change so it shouldn't be too bad. May revise in the future.

Also wired up the transaction metadata calls in preparation for working on the transaction object itself

Few notes in the README as a reminder to self