Updated README, Renamed Client to MonzoClient, Added Initial Transaction Support
Everything revolves around the client class, the name was ambiguous and could easily clash with other frameworks and I didn't want it to get confusing. I've decided to rename that to MonzoClient - everything else will keep it's name and take advantage of namespacing when there actually is a collision.

Initial work has been done for transactions, and the ping event has also been wired up.